Fixed panel height issue during the back animation
When hiding a panel with few content (viewport is not tall enough to cover the entire screen), the height property is removed too soon that causes panel to appear not full-height while animating away.

Now remove viewport and container height when panel is closed (when animation is finished).

Also changed scrolling to 0,0 instead of 0,1 coordinates when panel is closed that fixes noticeable content 1px scroll and preserves hiding address bar (due to following ignored window.scroll to 0,1 on sidebar show [line 108]).